# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Автопрокат 777
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Car {
  id: number;
  name: string;
  /** sedan, crossover, minivan, convertible, economy */
  category: string;
  /** Price in rubles per day */
  pricePerDay: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  seats?: number | null;
  /** @nullable */
  transmission?: string | null;
  /** @nullable */
  year?: number | null;
  popular?: boolean;
  available: boolean;
}

export interface BookingInput {
  name: string;
  phone: string;
  /** @nullable */
  carId: number | null;
  /** @nullable */
  carName?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  city: string;
  /** @nullable */
  comment?: string | null;
  withDriver?: boolean;
}

export interface Booking {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  carId?: number | null;
  /** @nullable */
  carName?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  city: string;
  /** @nullable */
  comment?: string | null;
  withDriver?: boolean;
  createdAt: string;
}

export interface Review {
  id: number;
  author: string;
  text: string;
  rating: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Branch {
  id: number;
  name: string;
  address: string;
  phone: string;
  city?: string;
  /** @nullable */
  workingHours?: string | null;
  /** @nullable */
  mapUrl?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
}

export type ListCarsParams = {
category?: string;
};