# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Автопрокат 777
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List all cars
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
...
  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "404":
          description: Car not found
...
components:
  schemas:
...
    Car:
      type: object
      required: [id, name, category, pricePerDay, available]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "sedan, crossover, minivan, convertible, economy"
        pricePerDay:
          type: integer
          description: Price in rubles per day
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        seats:
          type: ["integer", "null"]
        transmission:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        popular:
          type: boolean
        available:
          type: boolean
...