# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Автопрокат 777
 * OpenAPI spec version: 0.1.0
 */

export interface Car {
  id: number;
  name: string;
  /** sedan, crossover, minivan, convertible, economy */
  category: string;
  /** Price in rubles per day */
  pricePerDay: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  seats?: number | null;
  /** @nullable */
  transmission?: string | null;
  /** @nullable */
  year?: number | null;
  popular?: boolean;
  available: boolean;
}